+++ resources/views/auth/create.edge
@@ -2,7 +2,7 @@
 
 @section('body')
 <h1>Login</h1>
-<form method="POST" action="{{ route('sessions/store')}}">
+<form method="POST" action="{{ route('auth/store')}}">
   <input type="email" name="email" placeholder="E-mail">
   <input type="password" name="password" placeholder="Senha">
   <button type="submit">Entrar</button>
