+++ start/routes.ts
@@ -21,24 +21,24 @@
 import Route from '@ioc:Adonis/Core/Route'
 
 Route.get('/', async ({ view }) => {
-  return view.render('home')
+  return view.render('index')
 })
 
-Route.get('/login', 'SessionsController.create')
-  .as('sessions/create')
-Route.post('/login', 'SessionsController.store')
-  .as('sessions/store')
-Route.get('/logout', 'SessionsController.destroy')
-  .as('sessions/destroy')
+Route.get('/login', 'AuthController.create')
+  .as('auth/create')
+Route.post('/login', 'AuthController.store')
+  .as('auth/store')
+Route.get('/logout', 'AuthController.destroy')
+  .as('auth/destroy')
 
 Route.get('/cadastro', 'UsersController.create')
   .as('user/create')
 Route.post('/cadastro', 'UsersController.store')
   .as('user/store')
+Route.get('/editar-perfil', 'UsersController.edit')
+  .as('user/edit')
+  .middleware('auth')
+Route.post('/editar-perfil', 'UsersController.update')
+  .as('user/update')
+  .middleware('auth')
 
-Route.get('/editar-perfil', 'EditProfilesController.create')
-  .as('editProfile/create')
-  .middleware('auth')
-Route.post('/editar-perfil', 'EditProfilesController.update')
-  .as('editProfile/update')
-  .middleware('auth')
